Only import dirac fs or scheduler when DIRAC is installed
