+++ src/bartender/shared/__init__.py
@@ -1,0 +1,1 @@
+"""Module for things shared between filesystems and schedulers."""
